[FEAT]/33: Lang icon is now with responsiveness
