+++ src/atoms/index.js
@@ -12,5 +12,5 @@ export default function install(app) {
   app.component('a-progress-bar', ProgressBar)
   app.component('a-resume-play', ResumePlay)
   app.component('a-thumbnails', Thumbnails)
-  app.component('a-logo', Logo)
+  app.component('a-logo', Logo)  
 }
